refactor: use typescript for api-handlers
